m7. t2. fixed: added data-adapter.js
