+++ js/game-logic.js
@@ -1,8 +1,7 @@
 import {getClone} from "./utils";
 import {
   AnswerTimeType,
-  AnswerType,
-  QuestionType} from "./game-config";
+  AnswerType} from "./game-config";
 
 export const AnswerScoreType = {
   CORRECT: 100,
@@ -15,15 +14,7 @@ export const GameLevel = {
   MIN: 1,
   MAX: 10,
 };
-export const ImageTypeMap = {
-  'painting': `paint`,
-  'photo': `photo`,
-};
 
-export const RightAnswerTypeMap = {
-  'рисунок': `painting`,
-  'фото': `photo`,
-};
 export const ResultType = {
   RIGHT_ANSWER: `right`,
   FAST_ANSWER: `fast`,
@@ -127,52 +118,4 @@ export const getGameResultTotal = (gameResults) => {
   }, 0);
 
   return totalPoints;
-};
-
-export const adaptServerQuestionsData = (serverQuestionsData) => {
-  return serverQuestionsData.map((question) => {
-    return adaptQuestionData(question);
-  });
-};
-
-const adaptQuestionData = (questionData) => {
-  return {
-    type: questionData.type,
-    description: questionData.question,
-    images: adaptQuestionImagesData(questionData.answers),
-    rightAnswer: RightAnswerDataTypeMap[questionData.type](questionData),
-  };
-};
-
-const adaptQuestionImagesData = (imagesData) => {
-  return imagesData.map((answer) => {
-    return {
-      url: answer.image.url,
-      type: ImageTypeMap[answer.type],
-    };
-  });
-};
-
-const adaptRightAnswerDataOfTypeOne = (questionData) => {
-  return ImageTypeMap[questionData.answers[0].type];
-};
-
-const adaptRightAnswerDataOfTypeTwo = (questionData) => {
-  return ImageTypeMap[questionData.answers[0].type] + `-` + ImageTypeMap[questionData.answers[1].type];
-};
-
-const adaptRightAnswerDataOfTypeThree = (questionData) => {
-  const currentAnswerType = questionData.question.split(` `)[1];
-
-  const rightAnswerData = questionData.answers.filter((answer) => {
-    return answer.type === RightAnswerTypeMap[currentAnswerType];
-  });
-
-  return rightAnswerData[0].image.url;
-};
-
-const RightAnswerDataTypeMap = {
-  [QuestionType.ONE_IMAGE]: adaptRightAnswerDataOfTypeOne,
-  [QuestionType.TWO_IMAGES]: adaptRightAnswerDataOfTypeTwo,
-  [QuestionType.THREE_IMAGES]: adaptRightAnswerDataOfTypeThree,
 };
